% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_data}
\alias{create_data}
\title{Generate a data frame with random age, sex, and group variables}
\usage{
create_data(n = 100, create_errors = FALSE)
}
\arguments{
\item{n}{An integer value indicating the number of observations to generate.}

\item{create_errors}{A logical value indicating whether to generate errors in the data.}
}
\value{
A data frame with the generated variables.
}
\description{
This function generates a data frame with random age, sex, and group variables.
}
\examples{
create_data()

}
