% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_agepyramide}
\alias{plot_agepyramide}
\title{Plot a population pyramid}
\usage{
plot_agepyramide(
  age = create_data()$age,
  sex = create_data()$sex,
  group = create_data()$group,
  agegroup = TRUE,
  agegroup_size = 10
)
}
\arguments{
\item{age}{A numeric vector of ages.}

\item{sex}{A character vector of sexes.}

\item{group}{A character vector of groups.}

\item{agegroup}{A logical value indicating whether to group ages into chunks of 10 years.}

\item{agegroup_size}{An integer value indicating the size of age groups.}
}
\value{
A ggplot object.
}
\description{
This function creates a population pyramid using the ggplot2 package.
}
\examples{
plot_agepyramide()

}
