% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_agepyramide}
\alias{plot_agepyramide}
\title{Plot a population pyramid}
\usage{
plot_agepyramide(
  age = create_data()$age,
  sex = create_data()$sex,
  agegroup = TRUE,
  agegroup_size = 10,
  filter_NA = TRUE
)
}
\arguments{
\item{age}{A numeric vector of ages.}

\item{sex}{A character vector of sexes.}

\item{agegroup}{A logical value indicating whether to group ages into chunks.}

\item{agegroup_size}{An integer value indicating the size of age groups. Defaults to 10.}

\item{filter_NA}{An logical value indicating wether to exclude NA values. Defaults to TRUE.}
}
\value{
A ggplot object.
}
\description{
This function creates a population pyramid using the ggplot2 package.
}
\examples{
plot_agepyramide()

}
